Refactor WebSocket URL generation: Move getWsUrl function to a separate utility file and update imports in Queue and BattleArena components.
